Modifed the latest_disposition view to represent all of the columns in the disposition table instead of being a secondary table in a relationship between description and disponsition. It now has a form. We now use it in an outer join to disposition when searching for disposition attributes, so only the disposition from the most recent planning period is used for searching and the report table. The left join allows searching for projects without any disposition. The original PPT did not allow that. This resolves issue #8.
Modified project filtering and the report table to allow for selecting projects with a null
value for some attribute, which could be a nullable column with a select widget (like
sponsor) or a column with a multiple select widget (like driver or strategy) or a column in
a table (like disposition or comment (not implemented here). The original PPT appears to do
the first but the breakdown numbers are wrong, and appears to do the second but the report
does not work and again the counts are wrong.